% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_win_ratio.R
\name{calculate_win_ratio}
\alias{calculate_win_ratio}
\title{calculate_win_ratio}
\usage{
calculate_win_ratio(wins, losses)
}
\arguments{
\item{wins}{Number of wins}

\item{losses}{Number of losses}
}
\value{
A float
}
\description{
calculate_win_ratio
}
