% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sports_plot.R
\name{create_sports_plot}
\alias{create_sports_plot}
\title{Title}
\usage{
create_sports_plot(data)
}
\arguments{
\item{data}{Give a dataframe with team win/loss value}
}
\value{
a ggplot with custom team
}
\description{
Title
}
\examples{
create_sports_plot(data)

}
